createMessage implemented with Front and error handling improved

+++ src/main/java/org/chatop/chatopback/dto/rental/RentalResponseDto.java
@@ -3,7 +3,6 @@ package org.chatop.chatopback.dto.rental;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.net.URL;
 import java.time.LocalDateTime;
 
@@ -14,8 +13,8 @@ public record RentalResponseDto(
 
         Integer id,
         String name,
-        BigDecimal surface,
-        BigDecimal price,
+        Long surface,
+        Long price,
         URL picture,
         String description,
         Integer ownerId,
